refactor: 회의 반영 디렉토리 구조 변경
