+++ src/store/slices/candidateSlice.ts
